requirements.txt
# FastAPI + DB
fastapi
uvicorn
pydantic
sqlalchemy
asyncpg
python-dotenv
sqlmodel

# HuggingFace + Quantization
transformers
accelerate
bitsandbytes
torch
